move logic from all Controllers into module Devise::Controllers::*Controller and use this new modules in the Controllers
